add Kolnovel, remove node-fetch
