+++ libs/fetch.ts
@@ -1,14 +1,9 @@
-import { NodeFetchParams, getNodeFetch } from "./nodeFetch";
-
-const fetch = async (...args: NodeFetchParams) =>
-    (await getNodeFetch())(...args);
-
 const defaultUserAgentString =
     "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36";
 
 export async function fetchApi(
     url: string,
-    init?: NodeFetchParams[1],
+    init?: RequestInit,
 ) {
     const headers = {
         "User-Agent": defaultUserAgentString,
@@ -17,7 +12,7 @@ export async function fetchApi(
     return await fetch(url, { ...init, headers });
 }
 
-export const fetchFile = async function (url: string, init: NodeFetchParams[1]) {
+export const fetchFile = async function (url: string, init?: RequestInit) {
     if (!init) init = {};
     try {
         const res = await fetch(url, init);
